resolvendo problemas

+++ node.c
@@ -13,6 +13,9 @@ Node *node_construct(data_type value, Node *next) {
 }
 
 void node_destroy(Node *n, void (*destroy_fn)(data_type)) {
-    destroy_fn(n->data);
-    free(n);
+    if (n != NULL) {
+        if (destroy_fn != NULL)
+            destroy_fn(n->data);
+        free(n);
+    }
 }
